Collect sanitization functions into a single file
