+++ src/test/unit/process-profile.test.js
@@ -8,8 +8,8 @@ import {
   processProfile,
   serializeProfile,
   unserializeProfileOfArbitraryFormat,
-  sanitizePII,
 } from '../../profile-logic/process-profile';
+import { sanitizePII } from '../../profile-logic/sanitize';
 import { UniqueStringArray } from '../../utils/unique-string-array';
 import {
   createGeckoProfile,
